DIRECTOR: Move MCI parser to lingo-mci.cpp
- Make some small formatting changes.
- Removed int cast from uint when indexing
